Fix stopped notification firing too early
Was overwriting a child pid that I needed... oops.

Not much of a point in waiting for stopping... to finish. But I do want
to wait for stopped just in case that's the last thing the thread does,
so it doesn't kill the child too early (if the user specified a long
running command for notify for some reason...)