Reloading is now a thing!
Change settings without having to bring down your entire server stack!
Depending on the setting you may not even have to restart a single
server.

Currently the only keys that will trigger a restart are user, group,
path, log, and run.

Default would have no reason to change anything *while the daemon is
running*, before only matters when a server is being started, not while
it is running, and as such does not require a restart - same with after.
Notify can easily be changed as it is run each time a notification
needs to be sent, and therefore does not require a restart.