Small, hacky update, that adds non-privileged...
So basically, someone can set this up on a non-root account and use
systemctl --user. It takes manual setup, and I will not be adding an
install script as everyone may want to set it up in a different way and
location.

There is an example user service file included. Replace the fake paths
with real ones.